New (& incompatible) screen size hack
This is based on a patch by Charles Perkins:
https://github.com/charlesap/io/commit/cf23cea018c3b469b18091818a61baf0eebf07f6

It's not compatible with the previous hack, so if you were using that
you'll have to update Display.Mod. A small change to Input.Mod is also
need if you want to increase the size of the display.

The normal, unmodified Oberon software still works correctly, of
course.