Add raw serial support on Windows
On Windows, named pipes work differently from POSIX. They are
bidirectional, but one side needs to create it and the other needs to
open it (similar to sockets).

Therefore, the semantics for --serial-in and --serial-out are slightly
different: When --serial-in is used, a pipe is created, and when
--serial-out is used, the emulator opens the pipe/device.

If the device name does not start with a backslash, `\\.\pipe\` is
prepended, so you can use `--serial-in mypipe` instead of
`--serial-in \\.\pipe\mypipe`. In case the backslash is present, the
device name is used as is, so you should be able to use other devices
like `\\.\com4` as well.

Tested with both PuTTY and another emulator at the other end of the
pipe.